+++ ReleaseNotes.md
@@ -1,7 +1,11 @@
 # LightVx - Release Notes
-20/02/2018 - Version 1.1.0
+08/03/2018 - Version 1.1.0
 
 * Added USStateValidator and Extensions for Fluent API - use namespace LightVx.Validators.US
+* Added Min and Max Date Validators
+* Added Min/Max/IsAfter and IsBefore date validation to the Fluent API
+* Additional unit test coverage and bug fixes 
+* AlphaNumberic and Aplha validators now allow empty string.
 
 20/02/2018 - Version 1.0.3